requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
allure-pytest==2.15.0
allure-python-commons==2.15.0
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.75.0
anyio==4.9.0
asgiref==3.9.1
attrs==25.3.0
Authlib==1.6.6
backoff==2.2.1
beautifulsoup4==4.14.3
billiard==4.2.1
browser-use==0.10.1
bubus==1.5.6
cachetools==6.2.4
cdp-use==1.4.4
celery==5.3.4
certifi==2025.7.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpickle==3.1.2
colorama==0.4.6
croniter==6.0.0
cryptography==41.0.7
distro==1.9.0
Django==4.2.7
django-cors-headers==4.3.1
django-filter==23.5
django-simpleui==2025.6.24
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
docstring_parser==0.17.0
drf-spectacular==0.27.0
email-validator==2.3.0
Faker==40.1.2
fakerx==0.2.0
frozenlist==1.8.0
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.45.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.2
google-genai==1.56.0
googleapis-common-protos==1.72.0
greenlet==3.3.0
groq==1.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.3
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
inquirerpy==0.3.4
jiter==0.12.0
jsonpatch==1.33
jsonpath-ng==1.7.0
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
langchain-core==1.2.6
langchain-openai==1.1.6
langsmith==0.6.0
lxml==6.0.0
markdown-it-py==4.0.0
markdownify==1.2.2
mcp==1.25.0
mdurl==0.1.2
multidict==6.7.0
oauthlib==3.3.1
ollama==0.6.1
openai==2.14.0
orjson==3.11.5
outcome==1.3.0.post0
packaging==25.0
pfzy==0.3.4
playwright==1.57.0
pluggy==1.6.0
ply==3.11
portalocker==2.10.1
posthog==7.4.2
prompt_toolkit==3.0.51
propcache==0.4.1
proto-plus==1.27.0
protobuf==6.33.2
psutil==7.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyee==13.0.0
Pygments==2.19.2
PyJWT==2.10.1
PyMySQL==1.1.0
pyotp==2.9.0
pyparsing==3.3.1
pypdf==6.5.0
PyPDF2==3.0.1
PySocks==1.7.1
pytest==8.4.1
python-barcode==0.16.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.21
pytz==2025.2
PyYAML==6.0.2
pyzbar==0.1.9
qrcode==8.2
redis==5.0.1
referencing==0.36.2
regex==2025.11.3
reportlab==4.2.0
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.26.0
rsa==4.9.1
screeninfo==0.8.1
selenium==4.15.2
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.1
sqlparse==0.5.3
sse-starlette==3.1.2
starlette==0.50.0
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
trio==0.32.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uuid7==0.1.0
uuid_utils==0.12.0
uvicorn==0.40.0
vine==5.1.0
wcwidth==0.2.13
webdriver-manager==4.0.1
websockets==15.0.1
wsproto==1.3.2
yarl==1.22.0
zstandard==0.25.0

# UI自动化测试依赖包
playwright>=1.40.0
selenium==4.15.2
webdriver-manager==4.0.1
browser-use==0.10.1

# JWT 安全机制
djangorestframework-simplejwt==5.5.1
PyJWT>=2.8.0

# 测试数据生成
fakerx>=0.1.0

# APP自动化测试依赖包
airtest==1.4.3
loguru==0.7.3
Pillow==12.1.0

# OCR 文字识别支持
easyocr==1.7.2
opencv-contrib-python==4.6.0.66
numpy==1.26.4

# websocket推送进度
daphne==4.2.1
channels==4.3.2
channels-redis==4.3.0
